+++ packages/ui/src/components/select/index.ts
@@ -1,5 +1,5 @@
 export { default as CoarSelect } from './CoarSelect.vue';
 export { default as CoarMultiSelect } from './CoarMultiSelect.vue';
 export { default as CoarTagSelect } from './CoarTagSelect.vue';
-export type { CoarSelectOption } from './types';
+export type { CoarSelectOption, CoarSelectSortGroups, CoarSelectSortOptions } from './types';
 export type { CoarSelectSize, CoarSelectAppearance } from './useSelectBase';